refactor: add folder infra to group interfaces to nextjs
