--- config/blog.ts
@@ -1,6 +1,0 @@
-export const defaultSiteName = 'Blog do Lucas Fernando';
-export const defaultTitle = 'Blog do Lucas Fernando - Um blog sobre desenvolvimento web';
-export const defaultDescription =
-  'Um blog sobre desenvolvimento web, com publicações diárias de artigos sobre programação, css e noticias.';
-export const defaultThumbnail = 'https://blog.lucasfernandodev.com.br/assets/preview.png' ?? '';
-export const defaultUrl = 'https://blog.lucasfernandodev.com.br';
