+++ src/components/Layout/Header/Navigation/index.tsx
@@ -1,5 +1,5 @@
 import { useRouter } from 'next/router';
-import Link from '../../../Utils/Link';
+import Link from '../../../../infra/Link';
 import style from './style.module.css';
 
 export function Navigation() {
