+++ src/components/PostCards/index.tsx
@@ -2,8 +2,8 @@ import style from './style.module.css';
 import { Clock } from '../../lib/icons';
 import { BlogPost } from '../../types/post';
 import Tags from '../Tags';
-import Image from '../Utils/Image';
-import Link from '../Utils/Link';
+import Image from '../../infra/Image';
+import Link from '../../infra/Link';
 import Summary from '../Utils/Summary';
 import { cloudinaryImages } from '../../services/cloudinaryImages';
 import dynamic from 'next/dynamic';
